+++ Locale.lua
@@ -37,6 +37,7 @@ L["About"]                 = "About"
 L["Bank"]                  = "Bank"
 L["BANK SETTINGS"]         = "BANK SETTINGS"
 L["STASH LIST"]            = "STASH LIST"
+L["GUILD BANK"]            = "GUILD BANK"
 L["Mail"]                  = "Mail"
 L["MAIL SETTINGS"]         = "MAIL SETTINGS"